Add files via upload
Bank Marketing campaigns of a Portuguese banking institution need to identify the factors that cause the customers to tend to take the subscription, as well as Bank Marketing campaigns of a Portuguese banking institution need to identify the reasons behind the customer which make them not take the subscription.
So I have analyzed the factors for the subscription and nonsubscription of customers. Using the ITP and NPV techniques.